requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
psycopg2-binary==2.9.10
pydantic==2.11.7
SQLAlchemy==2.0.41
uvicorn[standard]==0.35.0
websockets==15.0.1
python-dotenv